+++ src/lib/objects/mod.rs
@@ -15,8 +15,8 @@ pub trait MeshObject: private::MeshObject {
     fn color(&self) -> [f32; 3];
     fn set_color(&mut self, color: [f32; 3]);
 
-    fn position(&self) -> cgmath::Point3<isize>;
-    fn set_position(&mut self, position: cgmath::Point3<isize>);
+    fn position(&self) -> cgmath::Point3<i16>;
+    fn set_position(&mut self, position: cgmath::Point3<i16>);
 
     fn light(&self) -> Option<[f32; 4]>;
     fn set_light(&mut self, light: Option<[f32; 4]>);
